+++ imking-web/src/main/java/io/imking/common/domain/Attach.java
@@ -1,4 +1,4 @@
-package io.imking.reward.domain;
+package io.imking.common.domain;
 
 import java.util.Date;
 import lombok.AllArgsConstructor;
