+++ imking-web/src/main/java/io/imking/common/domain/AttachExample.java
@@ -1,4 +1,4 @@
-package io.imking.reward.domain;
+package io.imking.common.domain;
 
 import java.util.ArrayList;
 import java.util.Date;
