--- imking-web/src/main/java/io/imking/common/domain/User.java
@@ -1,434 +1,0 @@
-package io.imking.common.domain;
-
-import java.util.Date;
-
-public class User {
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.id
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private Long id;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.account
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private String account;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.pwd
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private String pwd;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.realname
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private String realname;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.nickname
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private String nickname;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.qq
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private String qq;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.email
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private String email;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.website
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private String website;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.dream
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private String dream;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.points
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private Integer points;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.credit_points
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private Integer creditPoints;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.balance
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private Long balance;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column t_user.create_time
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    private Date createTime;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.id
-     *
-     * @return the value of t_user.id
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public Long getId() {
-        return id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.id
-     *
-     * @param id the value for t_user.id
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.account
-     *
-     * @return the value of t_user.account
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public String getAccount() {
-        return account;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.account
-     *
-     * @param account the value for t_user.account
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setAccount(String account) {
-        this.account = account;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.pwd
-     *
-     * @return the value of t_user.pwd
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public String getPwd() {
-        return pwd;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.pwd
-     *
-     * @param pwd the value for t_user.pwd
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setPwd(String pwd) {
-        this.pwd = pwd;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.realname
-     *
-     * @return the value of t_user.realname
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public String getRealname() {
-        return realname;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.realname
-     *
-     * @param realname the value for t_user.realname
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setRealname(String realname) {
-        this.realname = realname;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.nickname
-     *
-     * @return the value of t_user.nickname
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public String getNickname() {
-        return nickname;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.nickname
-     *
-     * @param nickname the value for t_user.nickname
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setNickname(String nickname) {
-        this.nickname = nickname;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.qq
-     *
-     * @return the value of t_user.qq
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public String getQq() {
-        return qq;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.qq
-     *
-     * @param qq the value for t_user.qq
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setQq(String qq) {
-        this.qq = qq;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.email
-     *
-     * @return the value of t_user.email
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public String getEmail() {
-        return email;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.email
-     *
-     * @param email the value for t_user.email
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.website
-     *
-     * @return the value of t_user.website
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public String getWebsite() {
-        return website;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.website
-     *
-     * @param website the value for t_user.website
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setWebsite(String website) {
-        this.website = website;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.dream
-     *
-     * @return the value of t_user.dream
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public String getDream() {
-        return dream;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.dream
-     *
-     * @param dream the value for t_user.dream
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setDream(String dream) {
-        this.dream = dream;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.points
-     *
-     * @return the value of t_user.points
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public Integer getPoints() {
-        return points;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.points
-     *
-     * @param points the value for t_user.points
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setPoints(Integer points) {
-        this.points = points;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.credit_points
-     *
-     * @return the value of t_user.credit_points
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public Integer getCreditPoints() {
-        return creditPoints;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.credit_points
-     *
-     * @param creditPoints the value for t_user.credit_points
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setCreditPoints(Integer creditPoints) {
-        this.creditPoints = creditPoints;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.balance
-     *
-     * @return the value of t_user.balance
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public Long getBalance() {
-        return balance;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.balance
-     *
-     * @param balance the value for t_user.balance
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setBalance(Long balance) {
-        this.balance = balance;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column t_user.create_time
-     *
-     * @return the value of t_user.create_time
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public Date getCreateTime() {
-        return createTime;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column t_user.create_time
-     *
-     * @param createTime the value for t_user.create_time
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    public void setCreateTime(Date createTime) {
-        this.createTime = createTime;
-    }
-}
