+++ imking-web/src/main/java/io/imking/common/mapping/ImkUserAccountDetailMapper.java
@@ -1,4 +1,4 @@
-package io.imking.reward.mapping;
+package io.imking.common.mapping;
 
 
 import java.util.List;
@@ -16,8 +16,8 @@ import org.apache.ibatis.annotations.Update;
 import org.apache.ibatis.annotations.UpdateProvider;
 import org.apache.ibatis.type.JdbcType;
 
-import io.imking.reward.domain.ImkUserAccountDetail;
-import io.imking.reward.domain.ImkUserAccountDetailExample;
+import io.imking.common.domain.ImkUserAccountDetail;
+import io.imking.common.domain.ImkUserAccountDetailExample;
 
 @Mapper
 public interface ImkUserAccountDetailMapper {
