--- imking-web/src/main/java/io/imking/common/mapping/UserMapper.java
@@ -1,130 +1,0 @@
-package io.imking.common.mapping;
-
-import org.apache.ibatis.annotations.Delete;
-import org.apache.ibatis.annotations.Insert;
-import org.apache.ibatis.annotations.InsertProvider;
-import org.apache.ibatis.annotations.Result;
-import org.apache.ibatis.annotations.Results;
-import org.apache.ibatis.annotations.Select;
-import org.apache.ibatis.annotations.Update;
-import org.apache.ibatis.annotations.UpdateProvider;
-import org.apache.ibatis.type.JdbcType;
-
-import io.imking.common.domain.User;
-
-public interface UserMapper {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    @Delete({
-        "delete from t_user",
-        "where id = #{id,jdbcType=BIGINT}"
-    })
-    int deleteByPrimaryKey(Long id);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    @Insert({
-        "insert into t_user (id, account, ",
-        "pwd, realname, nickname, ",
-        "qq, email, website, ",
-        "dream, points, credit_points, ",
-        "balance, create_time)",
-        "values (#{id,jdbcType=BIGINT}, #{account,jdbcType=VARCHAR}, ",
-        "#{pwd,jdbcType=VARCHAR}, #{realname,jdbcType=VARCHAR}, #{nickname,jdbcType=VARCHAR}, ",
-        "#{qq,jdbcType=VARCHAR}, #{email,jdbcType=VARCHAR}, #{website,jdbcType=VARCHAR}, ",
-        "#{dream,jdbcType=VARCHAR}, #{points,jdbcType=INTEGER}, #{creditPoints,jdbcType=INTEGER}, ",
-        "#{balance,jdbcType=BIGINT}, #{createTime,jdbcType=TIMESTAMP})"
-    })
-    int insert(User record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    @InsertProvider(type=UserSqlProvider.class, method="insertSelective")
-    int insertSelective(User record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    @Select({
-        "select",
-        "id, account, pwd, realname, nickname, qq, email, website, dream, points, credit_points, ",
-        "balance, create_time",
-        "from t_user",
-        "where id = #{id,jdbcType=BIGINT}"
-    })
-    @Results({
-        @Result(column="id", property="id", jdbcType=JdbcType.BIGINT, id=true),
-        @Result(column="account", property="account", jdbcType=JdbcType.VARCHAR),
-        @Result(column="pwd", property="pwd", jdbcType=JdbcType.VARCHAR),
-        @Result(column="realname", property="realname", jdbcType=JdbcType.VARCHAR),
-        @Result(column="nickname", property="nickname", jdbcType=JdbcType.VARCHAR),
-        @Result(column="qq", property="qq", jdbcType=JdbcType.VARCHAR),
-        @Result(column="email", property="email", jdbcType=JdbcType.VARCHAR),
-        @Result(column="website", property="website", jdbcType=JdbcType.VARCHAR),
-        @Result(column="dream", property="dream", jdbcType=JdbcType.VARCHAR),
-        @Result(column="points", property="points", jdbcType=JdbcType.INTEGER),
-        @Result(column="credit_points", property="creditPoints", jdbcType=JdbcType.INTEGER),
-        @Result(column="balance", property="balance", jdbcType=JdbcType.BIGINT),
-        @Result(column="create_time", property="createTime", jdbcType=JdbcType.TIMESTAMP)
-    })
-    User selectByPrimaryKey(Long id);
-
-    @Select({
-        "select",
-        "id,  nickname ",
-        "from t_user",
-        "where id = #{id,jdbcType=BIGINT}"
-    })
-    @Results({
-        @Result(column="id", property="id", jdbcType=JdbcType.BIGINT, id=true),
-        @Result(column="nickname", property="nickname", jdbcType=JdbcType.VARCHAR)
-    })
-    User selectNameByPrimaryKey(long id);
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    @UpdateProvider(type=UserSqlProvider.class, method="updateByPrimaryKeySelective")
-    int updateByPrimaryKeySelective(User record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table t_user
-     *
-     * @mbg.generated Wed Jan 17 21:43:23 CST 2018
-     */
-    @Update({
-        "update t_user",
-        "set account = #{account,jdbcType=VARCHAR},",
-          "pwd = #{pwd,jdbcType=VARCHAR},",
-          "realname = #{realname,jdbcType=VARCHAR},",
-          "nickname = #{nickname,jdbcType=VARCHAR},",
-          "qq = #{qq,jdbcType=VARCHAR},",
-          "email = #{email,jdbcType=VARCHAR},",
-          "website = #{website,jdbcType=VARCHAR},",
-          "dream = #{dream,jdbcType=VARCHAR},",
-          "points = #{points,jdbcType=INTEGER},",
-          "credit_points = #{creditPoints,jdbcType=INTEGER},",
-          "balance = #{balance,jdbcType=BIGINT},",
-          "create_time = #{createTime,jdbcType=TIMESTAMP}",
-        "where id = #{id,jdbcType=BIGINT}"
-    })
-    int updateByPrimaryKey(User record);
-}
