--- imking-web/src/main/java/io/imking/reward/domain/ImkUserPointDetail.java
@@ -1,25 +1,0 @@
-package io.imking.reward.domain;
-
-import java.util.Date;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ImkUserPointDetail {
-    private Integer id;
-
-    private Integer userId;
-
-    private Integer amount;
-
-    private Integer balance;
-
-    private Date createTime;
-
-    private String remark;
-}
