+++ imking-web/src/main/java/io/imking/reward/services/RewardService.java
@@ -4,17 +4,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.WeakHashMap;
 import java.util.stream.Collectors;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
-
-import io.imking.common.domain.User;
-import io.imking.common.mapping.UserMapper;
+import io.imking.common.mapping.ImkUserMapper;
 import io.imking.reward.beans.dto.RwDetailDto;
 import io.imking.reward.beans.enums.RewardStatusEnum;
 import io.imking.reward.domain.RwAnswerExample;
@@ -45,7 +41,7 @@ public class RewardService {
 	@Autowired
 	protected RwCommentMapper rwCommentMapper;
 	@Autowired
-	protected UserMapper userMapper;
+	protected ImkUserMapper userMapper;
 	@Autowired
 	protected RwAnswerMapper rwAnswerMapper;
 	/**
@@ -102,8 +98,8 @@ public class RewardService {
 			dto.setId(rwAsk.getId());
 			dto.setTitle(rwAsk.getTitle());
 			dto.setType(rwAsk.getType());
-			User user = userMapper.selectNameByPrimaryKey(rwAsk.getCreateBy().longValue());
-			dto.setNickname(user.getNickname());
+//			User user = userMapper.selectNameByPrimaryKey(rwAsk.getCreateBy().longValue());
+//			dto.setNickname(user.getNickname());
 			//rwAsk.get
 			rwAnswerExample
 				.createCriteria()
